Add exception cases, merge into main
